Refactor api folder structure and update imports
