+++ src/api/auth/index.ts
@@ -1,81 +1,8 @@
 import axios from 'axios';
-import {
-  CreateExerciseRequest,
-  DeleteExerciseRequest,
-  ExerciseApi,
-  ExerciseCreate,
-  Training,
-  TrainingApi,
-  UpdateExerciseRequest,
-  User,
-  UserApi,
-  UserMeasurement,
-  UserMeasurementApi,
-} from 'src/client/src';
+import { User, UserApi } from 'src/client/src';
 import { createInitOverrides } from '@utils/createInitOverrides';
 
-const trainingApi = new TrainingApi();
 const userApi = new UserApi();
-const MeasApi = new UserMeasurementApi();
-const ExerApi = new ExerciseApi();
-
-export const loadTrainings = async (
-  token: string,
-  page: number,
-  size: number,
-): Promise<Training[]> => {
-  const initOverrides = createInitOverrides(token);
-
-  const requestParameters = {
-    page: page,
-    size: size,
-  };
-  return await trainingApi.getTrainings(requestParameters, initOverrides);
-};
-
-export const addTraining = async (token: string, training: Training) => {
-  const initOverrides = createInitOverrides(token);
-
-  const requestParameters = {
-    trainingCreate: training,
-  };
-  return await trainingApi.createTraining(requestParameters, initOverrides);
-};
-
-export const updateTraining = async ({
-  token,
-  trainingId,
-  name,
-  description,
-}: {
-  token: string;
-  trainingId: string;
-  name: string;
-  description: string;
-}) => {
-  const initOverrides = createInitOverrides(token);
-
-  const requestParameters = {
-    trainingId: trainingId,
-    trainingCreate: { name, description },
-  };
-  return await trainingApi.updateTraining(requestParameters, initOverrides);
-};
-
-export const deleteTraining = async ({
-  token,
-  trainingId,
-}: {
-  token: string;
-  trainingId: string;
-}) => {
-  const initOverrides = createInitOverrides(token);
-
-  const requestParameters = {
-    trainingId: trainingId,
-  };
-  return await trainingApi.deleteTraining(requestParameters, initOverrides);
-};
 
 export const loginUser = async (creds: any) => {
   const response = await axios.post('http://188.68.247.208:8080/auth/signin', {
@@ -106,89 +33,4 @@ export const requestPasswordReset = async (email: string) => {
     email,
   );
   return response.data.message;
-};
-
-export const getUserMeasurement = async (token: string) => {
-  const initOverrides = createInitOverrides(token);
-
-  const requestParameters = {
-    userMeasurementId: 'b7393418-f187-4ada-999e-2158a2e8ae02',
-  };
-
-  return await MeasApi.getUserMeasurementById(requestParameters, initOverrides);
-};
-
-export const updateUserMeasurement = async ({
-  token,
-  userMeasurement,
-}: {
-  token: string;
-  userMeasurement: UserMeasurement;
-}) => {
-  const initOverrides = createInitOverrides(token);
-
-  const requestParameters = {
-    userMeasurementId: 'b7393418-f187-4ada-999e-2158a2e8ae02',
-    userMeasurementCreate: userMeasurement,
-  };
-  return await MeasApi.updateUserMeasurement(requestParameters, initOverrides);
-};
-
-export const getTrainingDetails = async (token: string, exerciseId: string) => {
-  const initOverrides = createInitOverrides(token);
-
-  const requestParameters = {
-    trainingId: exerciseId,
-  };
-  return await trainingApi.getTrainingById(requestParameters, initOverrides);
-};
-
-export const addExercise = async ({
-  token,
-  exerciseCreate,
-}: {
-  token: string;
-  exerciseCreate: ExerciseCreate;
-}) => {
-  const initOverrides = createInitOverrides(token);
-
-  const requestParameters: CreateExerciseRequest = {
-    exerciseCreate: exerciseCreate,
-  };
-  console.log('request add' + requestParameters);
-  return await ExerApi.createExercise(requestParameters, initOverrides);
-};
-
-export const updateExercise = async ({
-  token,
-  exerciseCreate,
-  exerciseId,
-}: {
-  token: string;
-  exerciseCreate: ExerciseCreate;
-  exerciseId: string;
-}) => {
-  const initOverrides = createInitOverrides(token);
-
-  const requestParameters: UpdateExerciseRequest = {
-    exerciseCreate: exerciseCreate,
-    exerciseId: exerciseId,
-  };
-  console.log('request update' + requestParameters);
-  return await ExerApi.updateExercise(requestParameters, initOverrides);
-};
-
-export const deleteExercise = async ({
-  token,
-  exerciseId,
-}: {
-  token: string;
-  exerciseId: string;
-}) => {
-  const initOverrides = createInitOverrides(token);
-  const requestParameters: DeleteExerciseRequest = {
-    exerciseId: exerciseId,
-  };
-
-  return await ExerApi.deleteExercise(requestParameters, initOverrides);
 };
